Отделяет templates
