+++ src/components/stats.js
@@ -1,19 +1,13 @@
 import {createElement} from '../utils.js';
+import {createStatsTemplate} from './../templates/stats-template.js';
 
-const createMovieListTemplate = () => {
-  return `<section class="films-list">
-    <h2 class="films-list__title visually-hidden">All movies. Upcoming</h2>
-    <div class="films-list__container"></div>
-  </section>`;
-};
-
-export default class MovieList {
+export default class Stats {
   constructor() {
     this._element = null;
   }
 
   getTemplate() {
-    return createMovieListTemplate();
+    return createStatsTemplate();
   }
 
   getElement() {
