+++ src/components/top-rated.js
@@ -1,11 +1,5 @@
 import {createElement} from '../utils.js';
-
-const createTopRatedTemplate = () => {
-  return `<section class="films-list--extra">
-    <h2 class="films-list__title">Top rated</h2>
-    <div class="films-list__container"></div>
-  </section>`;
-};
+import {createTopRatedTemplate} from './../templates/top-rated-template.js';
 
 export default class TopRated {
   constructor() {
