+++ src/templates/comment-template.js
@@ -1,6 +1,4 @@
-import {createElement} from '../utils.js';
-
-const createCommentTemplate = (comment) => {
+export const createCommentTemplate = (comment) => {
   const {emoji, text, author, date} = comment;
 
   return `<li class="film-details__comment">
@@ -17,26 +15,3 @@ const createCommentTemplate = (comment) => {
   </div>
   </li>`;
 };
-
-export default class Comment {
-  constructor(comment) {
-    this._comment = comment;
-    this._element = null;
-  }
-
-  getTemplate() {
-    return createCommentTemplate(this._comment);
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
-  }
-}
