+++ src/templates/main-movie-list-template.js
@@ -1,0 +1,6 @@
+export const createMovieListTemplate = () => {
+  return `<section class="films-list">
+    <h2 class="films-list__title visually-hidden">All movies. Upcoming</h2>
+    <div class="films-list__container"></div>
+  </section>`;
+};
