+++ src/templates/menu-template.js
@@ -1,6 +1,4 @@
-import {createElement} from '../utils.js';
-
-const createMenuMarkup = (menuItem) => {
+export const createMenuMarkup = (menuItem) => {
   const {name, count} = menuItem;
 
   const countMark = createMenuCount(count);
@@ -15,7 +13,7 @@ const createMenuCount = (count) => {
   return `<span class="main-navigation__item-count">${count}</span>`;
 };
 
-const createMenuTemplate = (menuItems) => {
+export const createMenuTemplate = (menuItems) => {
   const menuMarkup = menuItems.map((it) => createMenuMarkup(it)).join(`\n`);
   return `<nav class="main-navigation">
     <div class="main-navigation__items">
@@ -24,26 +22,3 @@ const createMenuTemplate = (menuItems) => {
     <a href="#stats" class="main-navigation__additional">Stats</a>
   </nav>`;
 };
-
-export default class Menu {
-  constructor(menuItems) {
-    this._menuItems = menuItems;
-    this._element = null;
-  }
-
-  getTemplate() {
-    return createMenuTemplate(this._menuItems);
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
-  }
-}
