+++ src/templates/movie-count-template.js
@@ -1,0 +1,8 @@
+import {getRandomIntegerNumber} from '../utils.js';
+
+export const createMovieCountTemplate = () => {
+  const moviesCount = new Intl.NumberFormat(`ru-RU`).format((getRandomIntegerNumber(10000, 300000)));
+  return `<section class="footer__statistics">
+    <p>${moviesCount} movies inside</p>
+  </section>`;
+};
