+++ src/templates/movies-section-template.js
@@ -1,0 +1,3 @@
+export const createMoviesSectionTemplate = () => {
+  return `<section class="films"></section>`;
+};
