+++ src/templates/no-movies-template.js
@@ -1,0 +1,5 @@
+export const createNoMoviesTemplate = () => {
+  return `<section class="films-list">
+  <h2 class="films-list__title">There are no movies in our database</h2>
+  </section>`;
+};
